Allocate only the needed size for the format structure array
Since we can know "immediately" how many format structures are needed, we
can allocate immediately the correct size and zero the structure in one
go. This avoids using realloc which can invalidate existing pointers.

The worst case for format elements is:
  <string>%<nsca formatter><string>
so 3 times the number of '%' (not considering that the nsca formatter can
contain '%' too for example for time specs).

However, this does not work since format_parse is adding a new format
element for each character it finds in the original format string.

Change-Id: I0a031d4b178ff16b7ee3083ce21c6ac08b4d8521