Allocate only the needed size for the format structure array
Since we can know how many format structures are needed before
allocating the memory, we can allocate immediately the correct
size and zero the structure in one go.

This avoids using realloc which can invalidate existing pointers.

The worst case for format elements is:
  <string> (%<ncsa formatter><string>)*
so we can use 2 * (1 + number of '%' in string) to be sure we never have
less than necessary elements in the array. (Note that the ncsa formatter
can contain '%' so we will allocate more memory than necessary).

Change-Id: I0a031d4b178ff16b7ee3083ce21c6ac08b4d8521